+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [3.7.0](https://github.com/josias-r/boarding.js/compare/v3.6.1...v3.7.0) (2025-11-28)
+
+
+### Features
+
+* support rtl keyboard control navigation ([fd07a86](https://github.com/josias-r/boarding.js/commit/fd07a86d953d0300529d325e30282aeae7207b21))
+
+## [3.6.1](https://github.com/josias-r/boarding.js/compare/v3.6.0...v3.6.1) (2025-11-28)
+
+
+### Bug Fixes
+
+* correct typo in spelling of 'preferredSide' in popover options with backwards compatibility (PR [#30](https://github.com/josias-r/boarding.js/issues/30)) ([1bad8bd](https://github.com/josias-r/boarding.js/commit/1bad8bd708d82eb28b90c24b685016882e725df0))
+
 # [3.6.0](https://github.com/josias-r/boarding.js/compare/v3.5.2...v3.6.0) (2024-09-16)